+++ stories/components/Alert.stories.mdx
@@ -5,6 +5,12 @@ import Alert from '@shell/components/Alert';
 <Meta 
   title="Components/Alert" 
   component={Alert}
+  argTypes={{
+    status: {
+      options: ['success', 'warning', 'info', 'error'],
+      control: { type: 'select' }
+    }
+  }}
 />
 
 export const Template = (args, { argTypes }) => ({
@@ -18,14 +24,13 @@ export const Template = (args, { argTypes }) => ({
 
 <Canvas>
   <Story
-    name="Success"
+    name="Default"
     args={{
       message: 'Hello world',
-      status: 'warning',
+      status: 'success',
     }}>
     {Template.bind({})}
   </Story>
- 
 </Canvas>
 
 ### Import
