+++ stories/components/ButtonDropdown.stories.mdx
@@ -13,15 +13,22 @@ export const Template = (args, { argTypes }) => ({
   template:   '<ButtonDropdown v-bind="$props" />',
 })
 
-# Alert
+# Button Dropdown
 
+Dropdowns are toggleable lists of links or actions which are displayed by clicking on them.
+
+### Description
+
+- They are generic and designed for a variety of situations and structures.
+- Use dropdowns in tables to provide more options on rows if the space is limited.
+- Don't include call-to-action buttons in a dropdown. They should always be presented on their own as their goal is to capture the user's attention.
 
 <Canvas>
   <Story
-    name="ButtonDrop"
+    name="Default"
     args={{
-      buttonLabel: 'Select',
-      dropdownOptions: ['0', '1', '2', '5']
+      buttonLabel: 'Action',
+      dropdownOptions: ['New', 'Edit', 'Delete']
     }}>
     {Template.bind({})}
   </Story>
@@ -42,3 +49,15 @@ export const Template = (args, { argTypes }) => ({
 
 <ArgsTable of={ButtonDropdown} />
 
+
+### Accessibility
+
+Dropdown buttons toggle the visualisation of contextual menus. We've defined a set of good accessibility practices for
+dropdown buttons, which will help users with assistive technologies.
+
+- When a button has focus, pressing the #[code spacebar] and #[code enter] keys should open the dropdown menu.
+- Users should be able to navigate the different menu items by using the #[code up] / #[code down] arrows.
+- The dropdown menu should close by using the #[code escape] key. On closing, the focus is automatically repositioned onto the button that opened the menu.
+- It should have an #[code aria-expanded] attribute with a #[code true]/#[code false] value, to indicates the current state of the dropdown popup.
+- It should have an #[code aria-haspopup] attribute set to #[code true] to indicate the availability and type of interactive dropdown popup.
+- If the button does not have any accessible label or text for eg. icon-only dropdown buttons, use the #[code aria-label] attribute.
